move Pattern into its own file
