+++ crates/matcher/src/lib.rs
@@ -1,52 +1,17 @@
 #[cfg(test)]
 mod tests;
 
+mod pattern;
+
 use cir::CIR;
-use trie_rs::map::{Trie, TrieBuilder};
+use trie_rs::map::Trie;
 
-pub trait Pattern {
-    fn pattern(&self) -> &[CIR];
-}
+pub use pattern::{Pattern, Patterns};
 
-pub struct Patterns<V> {
-    inner: TrieBuilder<cir::CIRKind, V>,
-}
-
-impl<V> Patterns<V> {
-    pub fn new() -> Self {
-        Self {
-            inner: TrieBuilder::new(),
-        }
-    }
-
-    pub fn finish(self) -> Matcher<V> {
-        Matcher {
-            inner: self.inner.build(),
-        }
-    }
-
-    pub fn push(&mut self, mut pattern: V, cir: impl FnOnce(&mut V) -> &[CIR]) {
-        let cir = cir(&mut pattern);
-        self.inner
-            .push(cir.iter().map(CIR::kind).collect::<Vec<_>>(), pattern);
-    }
-}
-
-impl<V> Default for Patterns<V> {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-pub struct Matcher<V> {
-    inner: Trie<cir::CIRKind, V>,
-}
-
-impl<V> Matcher<V> {
-    pub fn find_match(&self, cir: &[CIR]) -> Option<&V> {
-        self.inner
-            .exact_match(cir.iter().map(CIR::kind).collect::<Vec<_>>())
-    }
+#[derive(Debug)]
+pub struct Match<'a, 'b, V> {
+    value: &'a V,
+    matched: &'b [CIR],
 }
 
 /// Convenience function that wraps matched values together.
@@ -61,10 +26,15 @@ pub fn match_pair<'a, 'b, V>(
     })
 }
 
-#[derive(Debug)]
-pub struct Match<'a, 'b, V> {
-    value: &'a V,
-    matched: &'b [CIR],
+pub struct Matcher<V> {
+    inner: Trie<cir::CIRKind, V>,
+}
+
+impl<V> Matcher<V> {
+    pub fn find_match(&self, cir: &[CIR]) -> Option<&V> {
+        self.inner
+            .exact_match(cir.iter().map(CIR::kind).collect::<Vec<_>>())
+    }
 }
 
 impl<'a, 'b, V> Match<'a, 'b, V> {
